Refactor joinProject function: improve error handling, comments, and imports organization
Related to #1523
